third_party/WebKit/LayoutTests/http/tests/plugins/local-geturl-from-remote-expected.txt

CONSOLE ERROR: Not allowed to load local resource: tmp.html
This tests that a plug-in in a remote document can't access local files using NPN_GetURL

FAILURE
